update req to mongoDb
