+++ controllers/contacts.controller.js
@@ -1,8 +1,9 @@
-const contactsDb = require("../models/contacts");
+// const contactsDb = require("../models/contactsDb");
 const { HttpError } = require("../routes/helpers/index");
+const { Contacts } = require("../models/contacts");
 
 async function getContacts(req, res, next) {
-  const contacts = await contactsDb.listContacts();
+  const contacts = await Contacts.find({});
   if (!contacts) {
     return next(HttpError(404, "No contacts"));
   }
@@ -11,7 +12,7 @@ async function getContacts(req, res, next) {
 
 async function getContact(req, res, next) {
   const { contactId } = req.params;
-  const contact = await contactsDb.getContactById(contactId);
+  const contact = await Contacts.findById(contactId);
   if (!contact) {
     next(HttpError(404, "Contact not found"));
   }
@@ -20,17 +21,17 @@ async function getContact(req, res, next) {
 
 async function createContact(req, res, next) {
   const { name, email, phone } = req.body;
-  const newContact = await contactsDb.addContact(name, email, phone);
+  const newContact = await Contacts.create({ name, email, phone });
   res.status(201).json(newContact);
 }
 
 async function deleteContact(req, res, next) {
   const { contactId } = req.params;
-  const contact = await contactsDb.getContactById(contactId);
+  const contact = await Contacts.findById(contactId);
   if (!contact) {
     return next(HttpError(404, "No contact"));
   }
-  await contactsDb.removeContact(contactId);
+  await Contacts.findByIdAndRemove(contactId);
   return res.status(200).json({ message: "contact deleted" });
 }
 
@@ -41,11 +42,36 @@ async function updateContact(req, res, next) {
   if (Object.keys(body).length === 0) {
     return next(HttpError(400, "missing fields"));
   }
-  const contact = await contactsDb.getContactById(contactId);
+  const contact = await Contacts.findById(contactId);
   if (!contact) {
     return next(HttpError(404, "No contact"));
   }
-  const updatedContact = await contactsDb.updateContact(contactId, body);
+  const updatedContact = await Contacts.findByIdAndUpdate(
+    contactId,
+    { ...body },
+    { new: true }
+  );
+  res.status(200).json(updatedContact);
+}
+
+async function updateStatusContact(req, res, next) {
+  const { contactId } = req.params;
+
+  const { favorite } = req.body;
+  if (!favorite) {
+    return next(HttpError(400, "missing field favorite"));
+  }
+  const contact = await Contacts.findById(contactId);
+  if (!contact) {
+    return next(HttpError(404, "No contact"));
+  }
+  const updatedContact = await Contacts.findByIdAndUpdate(
+    contactId,
+    { favorite: favorite },
+    {
+      new: true,
+    }
+  );
   res.status(200).json(updatedContact);
 }
 
@@ -55,4 +81,5 @@ module.exports = {
   createContact,
   deleteContact,
   updateContact,
+  updateStatusContact,
 };
